+++ CHANGELOG.md
@@ -7,7 +7,9 @@ title: Changelog
 ## Unreleased (????-??-??)
 
 * Add check on `apiConfig` in cc-env-var-form.smart-exposed-config
-* `<cc-addon-linked-apps>`: use `<cc-zone>` instead of just zone name (BREAKING CHANGE)
+* `<cc-addon-linked-apps>`:
+  * Use `<cc-zone>` instead of just zone name (BREAKING CHANGE)
+  * Add smart definition
 
 ## 5.6.0 (2021-03-09)