+++ CHANGELOG.md
@@ -10,6 +10,7 @@ title: Changelog
 * `<cc-addon-linked-apps>`:
   * Use `<cc-zone>` instead of just zone name (BREAKING CHANGE)
   * Add smart definition
+* `<cc-env-var-form>`: add env-var-addon context
 
 ## 5.6.0 (2021-03-09)