+++ CHANGELOG.md
@@ -10,7 +10,9 @@ title: Changelog
 * `<cc-addon-linked-apps>`:
   * Use `<cc-zone>` instead of just zone name (BREAKING CHANGE)
   * Add smart definition
-* `<cc-env-var-form>`: add env-var-addon context
+* `<cc-env-var-form>`:
+  * add env-var-addon context
+  * add smart definition (env-var-addon)
 
 ## 5.6.0 (2021-03-09)