+++ CHANGELOG.md
@@ -6,10 +6,18 @@ title: Changelog
 
 ## Unreleased (????-??-??)
 
+...
+
+## 6.0.0 (2021-03-15)
+
+### ⚠️ BREAKING CHANGES
+
+* `<cc-addon-linked-apps>`: use `<cc-zone>` instead of just zone name (BREAKING CHANGE)
+
+### Components
+
 * Add check on `apiConfig` in cc-env-var-form.smart-exposed-config
-* `<cc-addon-linked-apps>`:
-  * Use `<cc-zone>` instead of just zone name (BREAKING CHANGE)
-  * Add smart definition
+* `<cc-addon-linked-apps>`: add smart definition
 * `<cc-env-var-form>`:
   * add env-var-addon context
   * add smart definition (env-var-addon)